requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-api-python-client==2.108.0
yt-dlp==2023.11.16
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
textblob==0.17.1
mlflow==2.8.1
python-dotenv==1.0.0
streamlit-option-menu==0.3.6
plotly==5.17.0 